pyproject.toml
[project]
name = "tshirts"
version = "0.1.0"
description = "A tool for breaking down GitHub issues into smaller pieces"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "pygithub>=2.1.0",
    "anthropic>=0.40.0",
    "rich>=13.0.0",
]

[project.scripts]
tshirts = "tshirts.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tshirts"]

[tool.hatch.build.targets.sdist]
include = ["src/tshirts"]
